simplify AerCircuit
