+++ qiskit_aer/backends/statevector_simulator.py
@@ -32,7 +32,6 @@ from .backend_utils import (cpp_execute_qobj, available_devices,
                             map_legacy_method_config,
                             add_final_save_op,
                             )
-from ..circuit.aer_circuit import generate_aer_circuits
 # pylint: disable=import-error, no-name-in-module
 from .controller_wrappers import aer_controller_execute
 
@@ -272,12 +271,11 @@ class StatevectorSimulator(AerBackend):
         qobj = map_legacy_method_options(qobj)
         return cpp_execute_qobj(self._controller, qobj)
 
-    def _execute_circuits(self, circuits, noise_model, config):
+    def _execute_circuits(self, aer_circuits, noise_model, config):
         """Execute circuits on the backend.
         """
         config = map_legacy_method_config(config)
-        aer_circuits = generate_aer_circuits(circuits)
-        circuits = add_final_save_op(aer_circuits, "statevector")
+        aer_circuits = add_final_save_op(aer_circuits, "statevector")
         return cpp_execute_circuits(self._controller, aer_circuits, noise_model, config)
 
     def _validate(self, qobj):
