+++ qiskit_aer/backends/unitary_simulator.py
@@ -33,7 +33,6 @@ from .backend_utils import (cpp_execute_qobj,
                             map_legacy_method_options,
                             add_final_save_op,
                             map_legacy_method_config)
-from ..circuit.aer_circuit import generate_aer_circuits
 # pylint: disable=import-error, no-name-in-module
 from .controller_wrappers import aer_controller_execute
 
@@ -271,12 +270,11 @@ class UnitarySimulator(AerBackend):
         qobj = map_legacy_method_options(qobj)
         return cpp_execute_qobj(self._controller, qobj)
 
-    def _execute_circuits(self, circuits, noise_model, config):
+    def _execute_circuits(self, aer_circuits, noise_model, config):
         """Execute circuits on the backend.
         """
         config = map_legacy_method_config(config)
-        aer_circuits = generate_aer_circuits(circuits)
-        circuits = add_final_save_op(aer_circuits, "unitary")
+        aer_circuits = add_final_save_op(aer_circuits, "unitary")
         return cpp_execute_circuits(self._controller, aer_circuits, noise_model, config)
 
     def _validate(self, qobj):
